backend/requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
motor==3.7.1
pydantic==2.12.5
python-dotenv==1.2.1
anthropic==0.75.0
httpx==0.28.1
google-generativeai==0.8.5
gunicorn==21.2.0